requirements.txt
certifi==2020.11.8
chardet==3.0.4
Django==2.2.16
django-appconf==1.0.4
django-debug-toolbar==2.2
django-image-optimizer==1.0.1
django-imagefit==0.6.2
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
python-resize-image==1.1.19
pytz==2020.1
requests==2.25.0
sqlparse==0.3.1
tinify==1.5.1
urllib3==1.26.2
